frontend: add create request page

+++ services/bidding-service/migrations/20221017316432_create_bidding_db.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS requests(
     Postcode VARCHAR(255),
     Info VARCHAR(255),
     CreatorId VARCHAR(255),
-    Deadline INTEGER,
+    Deadline BIGINT,
     Status VARCHAR(255)
 );
